requirements.txt
mysql-connector-python
langchain_core
dotenv
langchain_openai
langchain_groq
groq
langgraph
sqlparse